+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### [32.0.0] - Unreleased
 
+- Added
+  - Added flit build support. (GutHub #2)
+
 - Removed
   - Removed implicit release steps. (GitHub #6)